+++ docs/release_notes.rst
@@ -3,9 +3,10 @@
 Release notes
 =============
 
-* Added the ``GaussianForest`` and ``Fe55Forest`` model classes.
+* Added the ``GaussianForestBase`` and ``Fe55Forest`` model classes.
 * Pull requests merged and issues closed:
 
+  - https://github.com/lucabaldini/aptapy/pull/62
   - https://github.com/lucabaldini/aptapy/issues/47