+++ docs/release_notes.rst
@@ -3,6 +3,10 @@
 Release notes
 =============
 
+
+Version 0.16.0 (2025-12-03)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
 * Added the ``GaussianForestBase`` and ``Fe55Forest`` model classes.
 * Pull requests merged and issues closed: